% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{miRNAtap.db}
\alias{miRNAtap.db}
\alias{miRNAtap.db-package}
\title{database for miRNAtap package}
\description{
database for miRNAtap package
}
\details{
This package contains data annotations (target predictions).
}
\examples{
#direct targets in mouse aggregated from all sources:
library(miRNAtap)
targets_mouse <- getPredictedTargets('let-7a',species='mmu', method='geom')
}
\author{
Maciej Pajak \email{m.pajak@sms.ed.ac.uk}, Ian Simpson
}

